# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real-time chat application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  createdBy: string;
  createdAt: string;
  messageCount: number;
  /** @nullable */
  lastMessageAt?: string | null;
}

export interface RoomInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @maxLength 300 */
  description?: string;
}

export interface Message {
  id: number;
  roomId: number;
  userId: string;
  content: string;
  createdAt: string;
  user: AuthUser;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export interface RoomsSummary {
  totalRooms: number;
  totalMessages: number;
  onlineUsers: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};