# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real-time chat application API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface Message {
  id: number;
  roomId: number;
  userId: string;
  content: string;
  createdAt: Date;
  user: AuthUser;
}