# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real-time chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  createdBy: string;
  createdAt: Date;
  messageCount: number;
  /** @nullable */
  lastMessageAt?: Date | null;
}