# lib/api-zod/src/generated/types/roomsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real-time chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface RoomsSummary {
  totalRooms: number;
  totalMessages: number;
  onlineUsers: number;
}